Android-Only Heading Works like Google
Implemented the Android only Heading (without Google Play Services) to work like Google. Essentially, the right-side of the phone (when looking at it) points North and we subtract 90 degrees. That way, both the top of the phone and the back of the phone point to North, regardless of how the phone is held.

The plugin will still choose the Google Play Service Orientation Provider if it exists, since it is universally the same across devices.